force epoch timestamps

+++ backend/app/controllers/api/v1/capture_controller.rb
@@ -74,7 +74,7 @@ module Api
             uuid: e['uuid'] || "evt-#{SecureRandom.uuid}", 
             swishjam_api_key: public_key, 
             name: e['name'], 
-            occurred_at: Time.at(e['timestamp'] || Time.current.to_i),
+            occurred_at: Time.at(e['timestamp'] || Time.current.to_f),
             properties: (e['attributes'] || {}).as_json,
           )
         end
